// packages/protocol/typescript/generated/yiru/runtime/v1/git_common_pb.d.ts
export enum GitStatusArea {
  /**
   * @generated from enum value: GIT_STATUS_AREA_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: GIT_STATUS_AREA_STAGED = 1;
   */
  STAGED = 1,

  /**
   * @generated from enum value: GIT_STATUS_AREA_UNSTAGED = 2;
   */
  UNSTAGED = 2,

  /**
   * @generated from enum value: GIT_STATUS_AREA_UNTRACKED = 3;
   */
  UNTRACKED = 3,
}